Refactor recipe.controller.ts
There was a wrong thing in the controller. In some methods,
if the answer was not a truthy value, we would have a thrown
exception. Point is that, it would never be falsey value, and
mocking it could result in false positive test returns. In
order to correct it, that check has been removed.

Besides, allRecipe was redundant. If it had an empty object
as a parameter, it would return allRecipes value, otherwise
it would return filterRecipes value. Problem is that filter
method already returns allRecipes if an empty object is
passed. This was corrected by removing the conditional, and
returning filterRecipes value